+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+Increased max connections to 100. added retries to redis client.
+## [1.9.10] - 2026-02-11
+Set max connections to 50 for client
 ## [1.9.9] - 2025-09-28
 - Changed client retry
 ## [1.9.8] - 2025-08-21